Extract defaults into separated module
