+++ src/lib/class/OptionRegistry.js
@@ -1,33 +1,5 @@
 import Registry from './Registry';
-
-const defaults = {};
-
-defaults.defaultInvalidTypeProduct = null;
-defaults.defaultRandExpMax = 10;
-
-defaults.ignoreProperties = [];
-defaults.ignoreMissingRefs = false;
-defaults.failOnInvalidTypes = true;
-defaults.failOnInvalidFormat = true;
-
-defaults.alwaysFakeOptionals = false;
-defaults.optionalsProbability = false;
-defaults.fixedProbabilities = false;
-defaults.useExamplesValue = false;
-defaults.useDefaultValue = false;
-defaults.requiredOnly = false;
-
-defaults.minItems = 0;
-defaults.maxItems = null;
-defaults.minLength = 0;
-defaults.maxLength = null;
-
-defaults.resolveJsonPath = false;
-defaults.reuseProperties = false;
-defaults.fillProperties = true;
-defaults.replaceEmptyByRandomValue = false;
-
-defaults.random = Math.random;
+import defaults from '../api/defaults';
 
 /**
  * This class defines a registry for custom settings used within JSF.
